feat(spring) Changed JAXBContext to be in a producer pattern to avoid initing it every request.  I'm not sure if this is entirely the correct approach, however, but JAXBContext should be fine at least to init once (and this can be moved to per class instead of global)
Added back GETArtistAlbums

Added Rest Controller centered on Artist endpoints.  I'm not sure if this is the best approach to collect requests together, vs something similar to the initial pattern with Spark.

Despite the code for Storage only producing a NoneStorage, there exists an issue that Mongo is trying to connect anyway.  The MongoStorage is not the one being initialized, based on checking with the debugger.