feat(spring)
Abstracted the routes.  I'm not sure if this is the path that is good, vs having everything in one "repository" class, like other Spring tutorials have.  Some of the route classes do end up rather large, so having them all in one class may not be ideal in the long run (thus keeping the current separation of one class per operation)

Though, regardless, even if this path is the better of the two, I'm not sold on it myself.